some small changes, like removed couple of warnings; some minor changes in the test framework, added for comparison fw only and prepared statements to performance test. also made that test to turn timer on - so no need to run it w/ --timing option

+++ driver/mysql_connection.cpp
@@ -214,7 +214,7 @@ void MySQL_Connection::init(ConnectOptionsMap & properties)
 			}
 		} else if (!it->first.compare("port")) {
 			if ((p_ll = boost::get<long long>(&it->second))) {
-				port = *p_ll;
+        port = static_cast<unsigned int>(*p_ll);
 			} else {
 				throw sql::InvalidArgumentException("No long long value passed for port");
 			}
